ISSUES-4006 split connection packets
