+++ src/Core/MySQL/MySQLPackets.h
@@ -13,6 +13,19 @@ namespace DB
 namespace MySQLProtocol
 {
 
+enum Capability
+{
+    CLIENT_CONNECT_WITH_DB = 0x00000008,
+    CLIENT_PROTOCOL_41 = 0x00000200,
+    CLIENT_SSL = 0x00000800,
+    CLIENT_TRANSACTIONS = 0x00002000, // TODO
+    CLIENT_SESSION_TRACK = 0x00800000, // TODO
+    CLIENT_SECURE_CONNECTION = 0x00008000,
+    CLIENT_PLUGIN_AUTH_LENENC_CLIENT_DATA = 0x00200000,
+    CLIENT_PLUGIN_AUTH = 0x00080000,
+    CLIENT_DEPRECATE_EOF = 0x01000000,
+};
+
 class SSLRequest : public IMySQLReadPacket
 {
 public:
