+++ src/Core/MySQLProtocol.h
@@ -57,7 +57,6 @@ namespace MySQLProtocol
 
 //const size_t MAX_PACKET_LENGTH = (1 << 24) - 1; // 16 mb
 const size_t SCRAMBLE_LENGTH = 20;
-const size_t AUTH_PLUGIN_DATA_PART_1_LENGTH = 8;
 const size_t MYSQL_ERRMSG_SIZE = 512;
 const size_t PACKET_HEADER_SIZE = 4;
 const size_t SSL_REQUEST_PAYLOAD_SIZE = 32;
@@ -65,19 +64,6 @@ const size_t SSL_REQUEST_PAYLOAD_SIZE = 32;
 enum StatusFlags
 {
     SERVER_SESSION_STATE_CHANGED = 0x4000
-};
-
-enum Capability
-{
-    CLIENT_CONNECT_WITH_DB = 0x00000008,
-    CLIENT_PROTOCOL_41 = 0x00000200,
-    CLIENT_SSL = 0x00000800,
-    CLIENT_TRANSACTIONS = 0x00002000, // TODO
-    CLIENT_SESSION_TRACK = 0x00800000, // TODO
-    CLIENT_SECURE_CONNECTION = 0x00008000,
-    CLIENT_PLUGIN_AUTH_LENENC_CLIENT_DATA = 0x00200000,
-    CLIENT_PLUGIN_AUTH = 0x00080000,
-    CLIENT_DEPRECATE_EOF = 0x01000000,
 };
 
 enum Command
